requirements.txt
Flask
Gunicorn
open_clip
os
flask
flask_cors
numpy
torch
matplotlib.pyplot
torch.nn
PIL
collections
torchvision.datasets
torchvision